% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/AHEI2010-method.R
\docType{methods}
\name{AHEI2010}
\alias{AHEI2010}
\alias{AHEI2010,ffqr-method}
\title{AHEI-2010}
\usage{
AHEI2010(object)

\S4method{AHEI2010}{ffqr}(object)
}
\arguments{
\item{object}{a `ffqr` object}
}
\value{
a list of `tibbles`
}
\description{
Method to calculate the individual component scores for the Alternative Healthy Eating Index (AHEI) 2010 from a
Foor Frequency Questionaire (FFQ). The AHEI-2010 consists of 11 individual components which are all dietary components which are
linked to healthy eating and the incidence of chronic conditions.
}
