% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHEI_ComponentIndex.R
\docType{data}
\name{AHEI_ComponentIndex}
\alias{AHEI_ComponentIndex}
\title{AHEI2010 Component Index}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 64 rows and 2 columns.}
\usage{
AHEI_ComponentIndex
}
\description{
A dataset containing the food components needed for calculating the AHEI-2010
}
\keyword{datasets}
