% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahei_group_scoring.R
\name{ahei_group_scoring}
\alias{ahei_group_scoring}
\title{Score AHEI-2010 Component Groups}
\usage{
ahei_group_scoring(x, group)
}
\arguments{
\item{x}{a numeric value of the group raw value (ie, portions per day,grams per day, total mg, etc...)}

\item{group}{a character string indicating the group to score (`A1` - `A11`)}
}
\value{
a numeric value of the AHEI-2010 score for the specified group
}
\description{
Score AHEI-2010 Component Groups
}
