% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{AlcoholAHEI}
\alias{AlcoholAHEI}
\title{AHEI2010 Alcohol Correction}
\usage{
AlcoholAHEI(object, gender)
}
\arguments{
\item{object}{a `AHEI` object}

\item{gender}{a character string of Gender (F or M)}
}
\value{
a `AHEI` object
}
\description{
AHEI2010 Alcohol Correction
}
