% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/calculate_nutrients-method.R
\docType{methods}
\name{calculate_nutrients}
\alias{calculate_nutrients}
\alias{calculate_nutrients,ffqr-method}
\title{Calculate Nutrients}
\usage{
calculate_nutrients(object)

\S4method{calculate_nutrients}{ffqr}(object)
}
\arguments{
\item{object}{a `ffqr` object}
}
\value{
a `ffqr` object
}
\description{
Convert FFQ daily poritions to daily nutrient intake
}
