% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{check_value}
\alias{check_value}
\title{Check Item Value}
\usage{
check_value(x)
}
\arguments{
\item{x}{the `value` column of a FFQ `tibble`}
}
\value{
a numeric vector of position for removal
}
\description{
Check that each extracted variable is a food consumption catagoryu
}
\keyword{internal}
