% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_matching.R
\name{component_matching}
\alias{component_matching}
\title{Component Matching}
\usage{
component_matching(clean_ffq, index)
}
\arguments{
\item{clean_ffq}{a FFQ tibble which have been cleaned}

\item{index}{an index `tibble` containing a `FOOD_COMPONENT` column}
}
\value{
a `tibble` of matcing indicies
}
\description{
Match FFQ items into specified index tables (Portions & Nutrients)
}
\keyword{internal}
