% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{convert_servings}
\alias{convert_servings}
\title{Convert to Servings}
\usage{
convert_servings(x)
}
\arguments{
\item{x}{a FFQ response value}
}
\value{
a numeric value for the approimate servings per day
}
\description{
Convert FFQ responses to approximate servings per day
}
\keyword{internal}
