% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractA1.R
\name{ExtractA1}
\alias{ExtractA1}
\title{Extract FFQ items for AHEI A1}
\usage{
ExtractA1(x)
}
\arguments{
\item{x}{a `tibble` or `data.frame` of an FFQ}
}
\value{
a numeric value for the total servings per day of vegetables
}
\description{
Extract all the nesscary FFQ items which are used to calculate a score for AHEI group A1 (Vegetables)
}
