% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractA10.R
\name{ExtractA10}
\alias{ExtractA10}
\title{Extract FFQ items for AHEI A10}
\usage{
ExtractA10(x)
}
\arguments{
\item{x}{a `tibble` or `data.frame` of an FFQ}
}
\value{
a numeric value for the number of drinks per day
}
\description{
Extract all the nesscary FFQ items which are used to calculate a score for AHEI group A10 (Alcohol)
}
