% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractA2.R
\name{ExtractA2}
\alias{ExtractA2}
\title{Extract FFQ items for AHEI A2}
\usage{
ExtractA2(x)
}
\arguments{
\item{x}{a `tibble` or `data.frame` of an FFQ}
}
\value{
a numeric value for the total servings per day of fruit
}
\description{
Extract all the nesscary FFQ items which are used to calculate a score for AHEI group A2 (Fruit)
}
