% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractA3.R
\name{ExtractA3}
\alias{ExtractA3}
\title{Extract FFQ items for AHEI A3}
\usage{
ExtractA3(x)
}
\arguments{
\item{x}{a `tibble` or `data.frame` of an FFQ}
}
\value{
a numeric value for the total gramms per day of cereal fibre
}
\description{
Extract all the nesscary FFQ items which are used to calculate a score for AHEI group A3 (Cereal Fibre)
}
