% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractA4.R
\name{ExtractA4}
\alias{ExtractA4}
\title{Extract FFQ items for AHEI A4}
\usage{
ExtractA4(x)
}
\arguments{
\item{x}{a `tibble` or `data.frame` of an FFQ}
}
\value{
a numeric value for the total servings per day of sugar and sweetend drinks
}
\description{
Extract all the nesscary FFQ items which are used to calculate a score for AHEI group A4 (Sugar & Sweetend Drinks)
}
