% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractA5.R
\name{ExtractA5}
\alias{ExtractA5}
\title{Extract FFQ items for AHEI A5}
\usage{
ExtractA5(x)
}
\arguments{
\item{x}{a `tibble` or `data.frame` of an FFQ}
}
\value{
a numeric value for the total servings per day of nuts and legumes
}
\description{
Extract all the nesscary FFQ items which are used to calculate a score for AHEI group A5 (Nuts and Legumes)
}
