% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractA6.R
\name{ExtractA6}
\alias{ExtractA6}
\title{Extract FFQ items for AHEI A6}
\usage{
ExtractA6(x)
}
\arguments{
\item{x}{a `tibble` or `data.frame` of an FFQ}
}
\value{
a numeric value for the total servings per day of red and processed meat
}
\description{
Extract all the nesscary FFQ items which are used to calculate a score for AHEI group A6 (Red and Processed Meat)
}
