% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ffqr-class}
\alias{ffqr-class}
\title{FFQ Class}
\description{
A S4 class to store Food Frequency Questionnaires (FFQs)
}
\section{Slots}{

\describe{
\item{\code{FFQ}}{a list of cleaned FFQ tibbles}

\item{\code{Nutrients}}{a list of tibbles containing daily nutrient intake values for each FFQ food component}
}}

