% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nutrients.R
\docType{data}
\name{Nutrients}
\alias{Nutrients}
\title{FFQ Nutrients}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 148 rows and 43 columns.}
\usage{
Nutrients
}
\description{
A dataset containing amounts of nutrients in 100g of each food component
}
\keyword{datasets}
