% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openFFQ.R
\name{openFFQ}
\alias{openFFQ}
\title{Open Food Frequency Questionaire}
\usage{
openFFQ(ffq = list())
}
\arguments{
\item{ffq}{a list of Food Frequency Questionaires (FFQ)}
}
\value{
a S4 object (`ffqr`) which FFQ data converted to protions per day and grams per day
}
\description{
Open Food Frequency Questionaire
}
