% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portions.R
\docType{data}
\name{Portions}
\alias{Portions}
\title{FFQ Portions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 153 rows and 3 columns.}
\usage{
Portions
}
\description{
A dataset containing amounts of grams in a single portions
}
\keyword{datasets}
