% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_increments.R
\name{score_increments}
\alias{score_increments}
\title{Assign Incremental Scores}
\usage{
score_increments(start, end, n)
}
\arguments{
\item{start}{a numeric value where the scoring starts}

\item{end}{a numeric value where the scoring ends}

\item{n}{the number of groups that the increments are divided into}
}
\description{
Create incremental scores for each
}
\keyword{internal}
