% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/total_nutrients-method.R
\docType{methods}
\name{total_nutrients}
\alias{total_nutrients}
\alias{total_nutrients,ffqr-method}
\title{Total Nutrients}
\usage{
total_nutrients(object)

\S4method{total_nutrients}{ffqr}(object)
}
\arguments{
\item{object}{a `ffqr` object}
}
\value{
a list of `tibbles`
}
\description{
Extact the total nutrient profile from `ffqr` object
}
