% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/AlcoholAHEI-method.R
\docType{methods}
\name{AlcoholAHEI}
\alias{AlcoholAHEI}
\alias{AlcoholAHEI,ffqr-method}
\title{AHEI2010 Alcohol Correction}
\usage{
AlcoholAHEI(object, gender)

\S4method{AlcoholAHEI}{ffqr}(object, gender)
}
\arguments{
\item{object}{a `AHEI` object}

\item{gender}{a character string of Gender (F or M)}
}
\value{
a `AHEI` object
}
\description{

}
